refactor(bridge): move HiddenProjectsStore from routing/ to bridge/ directory
