+++ bridge/app/lib/src/bridge/hidden_projects_store.dart
