+++ bridge/app/lib/src/bridge/routing/close_project_handler.dart
@@ -1,6 +1,6 @@
 import "package:sesori_shared/sesori_shared.dart";
 
-import "hidden_projects_store.dart";
+import "../hidden_projects_store.dart";
 import "request_handler.dart";
 
 const _idParam = "id";
