+++ bridge/app/lib/src/bridge/routing/get_projects_handler.dart
@@ -3,7 +3,7 @@ import "dart:convert";
 import "package:sesori_plugin_interface/sesori_plugin_interface.dart";
 import "package:sesori_shared/sesori_shared.dart";
 
-import "hidden_projects_store.dart";
+import "../hidden_projects_store.dart";
 import "request_handler.dart";
 
 /// Handles `GET /project` — returns all projects from the plugin.
