+++ bridge/app/test/bridge/hidden_projects_store_test.dart
@@ -1,6 +1,6 @@
 import "dart:io";
 
-import "package:sesori_bridge/src/bridge/routing/hidden_projects_store.dart";
+import "package:sesori_bridge/src/bridge/hidden_projects_store.dart";
 import "package:test/test.dart";
 
 void main() {
