+++ bridge/app/test/bridge/routing/discover_project_handler_test.dart
@@ -1,8 +1,8 @@
 import "dart:convert";
 import "dart:io";
 
+import "package:sesori_bridge/src/bridge/hidden_projects_store.dart";
 import "package:sesori_bridge/src/bridge/routing/discover_project_handler.dart";
-import "package:sesori_bridge/src/bridge/routing/hidden_projects_store.dart";
 import "package:sesori_plugin_interface/sesori_plugin_interface.dart";
 import "package:test/test.dart";
