+++ bridge/app/test/bridge/routing/get_projects_handler_test.dart
@@ -1,8 +1,8 @@
 import "dart:convert";
 import "dart:io";
 
+import "package:sesori_bridge/src/bridge/hidden_projects_store.dart";
 import "package:sesori_bridge/src/bridge/routing/get_projects_handler.dart";
-import "package:sesori_bridge/src/bridge/routing/hidden_projects_store.dart";
 import "package:sesori_plugin_interface/sesori_plugin_interface.dart";
 import "package:test/test.dart";
